Refactorizada la exportación a ficheros txt
