+++ webapp/src/components/selectquestions/SelectQuestionsForm.jsx
@@ -3,6 +3,7 @@ import { DownloadOutlined, LeftOutlined } from "@ant-design/icons";
 import { useTranslation } from "react-i18next";
 import QuestionCardList from "./QuestionCardList";
 import { useState } from "react";
+import TxtExporter from "./export/txtexporter";
 
 const justifyButtonsBottom = {
   xs: "center",
@@ -47,20 +48,8 @@ function SelectQuestionsForm(props) {
 
   // Exportación a fichero de texto -------------------------------------------------
   const exportToTxt = () => {
-    let text = "";
-    questions.forEach((question) => {
-      if (question.question) {
-        text += question.question + "\r\n";
-        question.options.forEach((option) => { text += option.value +  (option.correct ? " T " : " F ") + "\r\n"; });
-      }
-      text += "\r\n";
-    });
-    const element = document.createElement("a");
-    const file = new Blob([text], { type: "text/plain" });
-    element.href = URL.createObjectURL(file);
-    element.download = "questions.txt";
-    document.body.appendChild(element); // Required for this to work in FireFox
-    element.click();
+    let txtExporter = new TxtExporter();
+    txtExporter.export(questions);
   }
 
   const { t } = useTranslation();
